chore(explorer): move CompactKeysTable props type to the table component
