+++ packages/apps/explorer/src/components/compact-keys-table/compact-keys-table-desktop/compact-keys-table-desktop.tsx
@@ -8,7 +8,7 @@ import {
 } from '@kadena/react-ui';
 import type { FC } from 'react';
 import React from 'react';
-import type { ICompactKeyTableProps } from '../types';
+import type { ICompactKeyTableProps } from '../compact-keys-table';
 import { tableClass } from './styles.css';
 
 const CompactTransactionsTableDesktop: FC<ICompactKeyTableProps> = ({
