+++ packages/apps/explorer/src/components/compact-keys-table/compact-keys-table.tsx
@@ -3,7 +3,16 @@ import type { FC } from 'react';
 import React from 'react';
 import CompactKeysTableDesktop from './compact-keys-table-desktop/compact-keys-table-desktop';
 import CompactKeysTableMobile from './compact-keys-table-mobile/compact-keys-table-mobile';
-import type { ICompactKeyTableProps } from './types';
+
+export interface IKeyProps {
+  chainId: string;
+  key: string;
+  predicate: string;
+}
+
+export interface ICompactKeyTableProps {
+  keys: IKeyProps[];
+}
 
 const CompactKeysTable: FC<ICompactKeyTableProps> = ({ keys }) => {
   return (
