--- packages/apps/explorer/src/components/compact-keys-table/types.ts
@@ -1,9 +1,0 @@
-export interface IKeyProps {
-  chainId: string;
-  key: string;
-  predicate: string;
-}
-
-export interface ICompactKeyTableProps {
-  keys: IKeyProps[];
-}
